Restructure directory tree. Fixed path in PJ. Updated all files for ./config/env path. All tests passing on local machine.
